* Fixed bug which was preventing motor from running in reverse. * Expanded time field in I2C mechanical brake command sent to   daughterboard from uint8_t to uint16_t so that we can actuate the brake   for longer than 255ms.

+++ commands.c
@@ -534,7 +534,7 @@ void cmdBrake(const char *args) {
 		return;
 	}
 
-	mechbrake_actuate(dir, (uint16_t)current_mA, (uint8_t)time_ms);
+	mechbrake_actuate(dir, (uint16_t)current_mA, (uint16_t)time_ms);
 }
 
 /****************/
